Split operations into a folder module
